+++ packages/terradart_google/lib/iap.dart
@@ -1,9 +1,15 @@
 // packages/terradart_google/lib/iap.dart
-/// Identity-Aware Proxy (IAP) IAM bindings for external HTTPS load balancer
-/// backend services.
+/// Identity-Aware Proxy (IAP) IAM for App Engine and external HTTPS load
+/// balancer backend services.
 library;
 
+export 'src/iap/google_iap_app_engine_service_iam_member.dart'
+    show GoogleIapAppEngineServiceIamMember;
+export 'src/iap/google_iap_app_engine_version_iam_member.dart'
+    show GoogleIapAppEngineVersionIamMember;
 export 'src/iap/google_iap_web_backend_service_iam_binding.dart'
     show GoogleIapWebBackendServiceIamBinding;
 export 'src/iap/google_iap_web_backend_service_iam_member.dart'
     show GoogleIapWebBackendServiceIamMember;
+export 'src/iap/google_iap_web_type_app_engine_iam_member.dart'
+    show GoogleIapWebTypeAppEngineIamMember;